sunset_fest_cabo_backend/requirements.txt
asgiref==3.8.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
PyJWT==2.9.0
python-dotenv==1.1.0
runserver==1.0.4
sqlparse==0.5.3
tzdata==2025.2
Werkzeug==3.1.3
stripe
requests
